requirements.txt
langchain>=0.2.10
langchain-openai>=0.1.20
streamlit>=1.36
python-dotenv>=1.0
langchain_community
langserve
fastapi
uvicorn
sse_starlette